parser/requirements.txt
psycopg2-binary>=2.9.10
openpyxl>=3.1.2
python-dotenv>=1.0.0